requirements.txt
Pillow==6.2.1
qrcode==6.1